@@ -1,14 +1,24 @@
 module github.com/corazawaf/coraza/v2
 
-go 1.16
+go 1.18
 
 require (
-	github.com/cloudflare/ahocorasick v0.0.0-20210425175752-730270c3e184
 	github.com/corazawaf/libinjection-go v0.0.0-20220207031228-44e9c4250eb5
 	github.com/foxcpp/go-mockdns v1.0.0
+	github.com/magefile/mage v1.13.0
+	github.com/petar-dambovaliev/aho-corasick v0.0.0-20211021192214-5ab2d9280aa9
+	go.uber.org/zap v1.22.0
+	golang.org/x/net v0.0.0-20220722155237-a158d28d115b
+	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/miekg/dns v1.1.50 // indirect
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
-	go.uber.org/zap v1.21.0
-	golang.org/x/net v0.0.0-20220520000938-2e3eb7b945c2
-	gopkg.in/yaml.v2 v2.4.0
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4 // indirect
+	golang.org/x/sys v0.0.0-20220808155132-1c4a2a72c664 // indirect
+	golang.org/x/tools v0.1.6-0.20210726203631-07bc1bf47fb2 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 )
