@@ -16,8 +16,8 @@ package seclang
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 
@@ -43,10 +43,12 @@ func Test_NonImplementedDirective(t *testing.T) {
 	w := coraza.NewWaf()
 	p, _ := NewParser(w)
 	for _, rule := range rules {
-		err := p.FromString(rule)
-		if err != nil {
-			t.Errorf("failed to set directive: %s", rule)
-		}
+		t.Run(rule, func(t *testing.T) {
+			err := p.FromString(rule)
+			if err != nil {
+				t.Errorf("failed to set directive: %s", rule)
+			}
+		})
 	}
 }
 
@@ -150,7 +152,7 @@ func Test_directive(t *testing.T) {
 
 func TestDebugDirectives(t *testing.T) {
 	waf := coraza.NewWaf()
-	tmpf, _ := ioutil.TempFile("/tmp", "*.log")
+	tmpf, _ := os.Create(filepath.Join(t.TempDir(), "debug.log"))
 	tmp := tmpf.Name()
 	p, _ := NewParser(waf)
 	err := directiveSecDebugLog(&DirectiveOptions{
@@ -204,7 +206,7 @@ func TestSecAuditLogDirectivesConcurrent(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	data, err := ioutil.ReadFile(f)
+	data, err := os.ReadFile(f)
 	if err != nil {
 		t.Error(err)
 	}
@@ -245,7 +247,7 @@ func TestSecRuleUpdateTargetBy(t *testing.T) {
 
 // Find a file by name recursively containing some string
 func findFileContaining(path string, search string) (string, error) {
-	files, err := ioutil.ReadDir(path)
+	files, err := os.ReadDir(path)
 	if err != nil {
 		return "", err
 	}
