@@ -19,11 +19,16 @@ import "testing"
 func TestDefaultWriters(t *testing.T) {
 	ws := []string{"serial", "concurrent"}
 	for _, writer := range ws {
-		if w, err := GetLogWriter(writer); err != nil {
-			t.Error(err)
-		} else if w == nil {
-			t.Errorf("invalid %s writer", writer)
-		}
+		t.Run(writer, func(t *testing.T) {
+			w, err := GetLogWriter(writer)
+			if err != nil {
+				t.Error(err)
+			}
+
+			if w == nil {
+				t.Errorf("invalid %q writer", writer)
+			}
+		})
 	}
 }
 func TestWriterPlugins(t *testing.T) {
