@@ -15,8 +15,9 @@
 package operators
 
 import (
-	"github.com/corazawaf/coraza/v2"
 	"github.com/corazawaf/libinjection-go"
+
+	"github.com/corazawaf/coraza/v2"
 )
 
 type detectSQLi struct {
