@@ -22,15 +22,20 @@ import (
 	"github.com/corazawaf/coraza/v2"
 )
 
+type byteRange struct {
+	start byte
+	end   byte
+}
+
 type validateByteRange struct {
-	data [][]byte
+	data []byteRange
 }
 
 func (o *validateByteRange) Init(data string) error {
-	o.data = [][]byte{}
 	if data == "" {
 		return nil
 	}
+	o.data = []byteRange{}
 	ranges := strings.Split(data, ",")
 	spl := ranges
 	var err error
@@ -71,19 +76,20 @@ func (o *validateByteRange) Evaluate(tx *coraza.Transaction, data string) bool {
 	if data == "" && lenData > 0 {
 		return false
 	}
-	input := []byte(data)
+
 	// we must iterate each byte from input and check if it is in the range
 	// if every byte is within the range we return false
 	matched := 0
-	for _, c := range input {
+	for i := 0; i < len(data); i++ {
+		c := data[i]
 		for _, r := range o.data {
-			if c >= r[0] && c <= r[1] {
+			if c >= r.start && c <= r.end {
 				matched++
 				break
 			}
 		}
 	}
-	return len(input) != matched
+	return len(data) != matched
 }
 
 func (o *validateByteRange) addRange(start uint64, end uint64) error {
@@ -93,6 +99,6 @@ func (o *validateByteRange) addRange(start uint64, end uint64) error {
 	if end > 255 {
 		return fmt.Errorf("invalid byte %d", end)
 	}
-	o.data = append(o.data, []byte{byte(start), byte(end)})
+	o.data = append(o.data, byteRange{byte(start), byte(end)})
 	return nil
 }
