@@ -26,8 +26,8 @@ import (
 type validateNidFunction = func(input string) bool
 
 type validateNid struct {
-	fn  validateNidFunction
-	rgx string
+	fn validateNidFunction
+	re *regexp.Regexp
 }
 
 func (o *validateNid) Init(data string) error {
@@ -43,13 +43,14 @@ func (o *validateNid) Init(data string) error {
 	default:
 		return fmt.Errorf("invalid @validateNid argument")
 	}
-	o.rgx = spl[1]
-	return nil
+
+	var err error
+	o.re, err = regexp.Compile(spl[1])
+	return err
 }
 
 func (o *validateNid) Evaluate(tx *coraza.Transaction, value string) bool {
-	re, _ := regexp.Compile(o.rgx)
-	matches := re.FindAllStringSubmatch(value, -1)
+	matches := o.re.FindAllStringSubmatch(value, -1)
 
 	res := false
 	for i, m := range matches {
