@@ -430,7 +430,7 @@ func getLastRuleExpectingChain(w *coraza.Waf) *coraza.Rule {
 
 // parseActions will assign the function name, arguments and
 // function (pkg.actions) for each action split by comma (,)
-// Action arguments are allowed to wrap values between colons('')
+// Action arguments are allowed to wrap values between colons(”)
 func parseActions(actions string) ([]ruleAction, error) {
 	iskey := true
 	ckey := ""
