@@ -14,44 +14,38 @@
 
 package transformations
 
+import "strings"
+
 func urlEncode(data string) (string, error) {
 	return doURLEncode(data), nil
 }
 
 func doURLEncode(input string) string {
 	inputLen := len(input)
-	var i, count, c int
-	leng := inputLen * 3
-	d := make([]byte, leng)
-	d = append(d, []byte(input)...)
+	if inputLen == 0 {
+		return ""
+	}
+
+	length := inputLen * 3
+	var d strings.Builder
+	d.Grow(length)
 	c2xTable := []byte("0123456789abcdef")
 
 	/* ENH Only encode the characters that really need to be encoded. */
 
-	for i = 0; i < inputLen; i++ {
+	for i := 0; i < inputLen; i++ {
 		cc := input[i]
 
 		if cc == ' ' {
-			d[c] = '+'
-			c++
-			count++
+			d.WriteByte('+')
 		} else {
 			if (cc == 42) || ((cc >= 48) && (cc <= 57)) || ((cc >= 65) && (cc <= 90)) || ((cc >= 97) && (cc <= 122)) {
-				d[c] = cc
-				c++
-				count++
+				d.WriteByte(cc)
 			} else {
-				d[c] = '%'
-				c++
-				count++
-				d[c] = c2xTable[(cc&0xff)>>4]
-				d[c+1] = c2xTable[(cc&0xff)&0x0f]
-				c += 2
-				count++
-				count++
+				d.Write([]byte{'%', c2xTable[(cc&0xff)>>4], c2xTable[(cc&0xff)&0x0f]})
 			}
 		}
 	}
 
-	return string(d[0:count])
+	return d.String()
 }
